+++ package.json
@@ -12,6 +12,7 @@
   "dependencies": {
     "@cloudinary/react": "^1.11.2",
     "@cloudinary/url-gen": "^1.11.2",
+    "@defer/client": "^1.12.1",
     "@prisma/client": "^5.4.1",
     "@react-email/components": "^0.0.7",
     "@stripe/stripe-js": "^2.1.6",
